Done logic in js files and organised php files

+++ public/backend/removedb.php
@@ -1,10 +1,20 @@
 <?php
-$database = new PDO('sqlite:your_database.db');
-$nhsNo = $_POST['nhsNo'];
-// Step 3: Create the SQL DELETE statement
-$delete_query = "DELETE FROM users WHERE nhsNo = :nhsNo";
-$statement = $database->prepare($delete_query);
-$statement->bindParam(':nhsNo', $nhsNo);
-$statement->execute();
-$database = null;
+
+try {
+
+        $db = new PDO('sqlite:surgery.db');
+        $db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+
+        $nhsNo = $_POST['nhsNumber'];
+        $password = $_POST['password'];
+
+    
+        if ($user) {
+            $delete_query = "DELETE FROM users WHERE nhsNo = :nhsNo";
+            $statement = $database->prepare($delete_query);
+            $statement->bindParam(':nhsNo', $nhsNo);
+            $statement->execute();
+            $database = null;
+}
+
 ?>
